+++ src/test/java/ladder/service/CustomLineStrategy.java
@@ -1,10 +1,9 @@
-package ladder.util;
+package ladder.service;
 
 import ladder.domain.Line;
 import ladder.domain.Step;
 import java.util.ArrayList;
 import java.util.List;
-import ladder.util.LineStrategy;
 
 public class CustomLineStrategy implements LineStrategy {
     private final List<Step> mockSteps;
