+++ src/test/java/ladder/service/LineStrategyTest.java
@@ -1,4 +1,4 @@
-package ladder.util;
+package ladder.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
